Changed the package structure to "/com/notify/notify" to follow common standards
