+++ app/build.gradle.kts
@@ -5,11 +5,11 @@ plugins {
 }
 
 android {
-    namespace = "com.example.test"
+    namespace = "com.notify.notify"
     compileSdk = 34
 
     defaultConfig {
-        applicationId = "com.example.test"
+        applicationId = "com.notify.notify"
         minSdk = 24
         targetSdk = 34
         versionCode = 1
