+++ app/src/androidTest/java/com/notify/notify/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.test
+package com.notify.notify
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
