--- app/src/main/java/com/example/test/screens/FirstFragment.kt
@@ -1,34 +1,0 @@
-package com.example.test.screens
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.test.MAIN
-import com.example.test.R
-import com.example.test.databinding.ActivityMainBinding
-import com.example.test.databinding.FragmentFirstBinding
-
-
-class FirstFragment : Fragment() {
-    lateinit var  binding: FragmentFirstBinding
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-      binding = FragmentFirstBinding.inflate(layoutInflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding.btnNext.setOnClickListener {
-            MAIN.navController.navigate(R.id.action_firstFragment_to_secondFragment)
-        }
-
-    }
-
-}
