+++ app/src/main/java/com/notify/notify/screens/FirstFragment.kt
@@ -1,30 +1,33 @@
-package com.example.test.screens
+package com.notify.notify.screens
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.test.MAIN
-import com.example.test.R
-import com.example.test.databinding.FragmentSecondBinding
+import com.notify.notify.MAIN
+import com.notify.notify.R
+import com.notify.notify.databinding.FragmentFirstBinding
 
 
-class SecondFragment : Fragment() {
+class FirstFragment : Fragment() {
+    lateinit var  binding: FragmentFirstBinding
 
-lateinit var binding: FragmentSecondBinding
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-       binding = FragmentSecondBinding.inflate(layoutInflater, container, false)
+      binding = FragmentFirstBinding.inflate(layoutInflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        binding.btnBack.setOnClickListener{
-            MAIN.navController.navigate(R.id.action_secondFragment_to_firstFragment)
+        binding.btnNext.setOnClickListener {
+            MAIN.navController.navigate(R.id.action_firstFragment_to_secondFragment)
         }
+
     }
+
 }
