+++ app/src/main/java/com/notify/notify/screens/SecondFragment.kt
@@ -1,0 +1,30 @@
+package com.notify.notify.screens
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.notify.notify.MAIN
+import com.notify.notify.R
+import com.notify.notify.databinding.FragmentSecondBinding
+
+
+class SecondFragment : Fragment() {
+
+lateinit var binding: FragmentSecondBinding
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+       binding = FragmentSecondBinding.inflate(layoutInflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        binding.btnBack.setOnClickListener{
+            MAIN.navController.navigate(R.id.action_secondFragment_to_firstFragment)
+        }
+    }
+}
