+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/firstFragment"
-        android:name="com.example.test.screens.FirstFragment"
+        android:name="com.notify.notify.screens.FirstFragment"
         android:label="fragment_first"
         tools:layout="@layout/fragment_first" >
         <action
@@ -16,7 +16,7 @@
     </fragment>
     <fragment
         android:id="@+id/secondFragment"
-        android:name="com.example.test.screens.SecondFragment"
+        android:name="com.notify.notify.screens.SecondFragment"
         android:label="fragment_second"
         tools:layout="@layout/fragment_second" >
         <action
